+++ include/cask/scheduler/ThreadPoolScheduler.hpp
@@ -35,41 +35,9 @@ public:
      */
     ~ThreadPoolScheduler();
 
-    /**
-     * Submit a task for execution in the thread pool. This task will
-     * execute after an indeterminite amount of time as resources free
-     * to perform the task.
-     * 
-     * @param task The task to submit for execution.
-     */
     void submit(const std::function<void()>& task) override;
-
-    /**
-     * Submit several tasks at once to the the thread pool. The order
-     * these tasks will be taken up and executed is undefined. Each
-     * task will execute after an indeterminite amount of time as
-     * resource free to perform the individual task.
-     * 
-     * @param tasks The vector of tasks to submit in-bulk.
-     */
     void submitBulk(const std::vector<std::function<void()>>& tasks) override;
-
-    /**
-     * Submit a task to the pool after _at least_ the given amount
-     * of time has passed.
-     * 
-     * @param milliseconds The number of milliseconds to wait before
-     *                     submitting to the pool
-     * @param task The task the submit after the wait time has elapsed.
-     */
     void submitAfter(int64_t milliseconds, const std::function<void()>& task) override;
-
-    /**
-     * Check if the scheduler is currently idle - meaning all threads are
-     * currently waiting for tasks to execute.
-     * 
-     * @return true if the scheduler is idle.
-     */
     bool isIdle() const override;
 private:
     bool running;